+++ src/vigipy/LongitudinalModel/LongitudinalModel.py
@@ -16,7 +16,7 @@ class LongitudinalModel:
         """
         self.time_unit = time_unit
         dataframe["date"] = pd.to_datetime(dataframe["date"])
-        self.date_groups = dataframe.resample(time_unit, on="date").sum()["count"]
+        self.date_groups = dataframe.resample(time_unit, on="date")
         self.data = dataframe
 
     def run(self, model, include_gaps=True, **kwargs):
@@ -35,10 +35,36 @@ class LongitudinalModel:
 
         """
         self.results = []
-        for timestamp, count in self.date_groups.iteritems():
+        for timestamp, count in self.date_groups.sum()["count"].iteritems():
             if count == 0 and not include_gaps:
                 continue
             subset = self.data.loc[self.data["date"] <= timestamp]
+            sub_container = convert(subset)
+            try:
+                da_results = model(sub_container, **kwargs)
+                self.results.append((timestamp, da_results))
+            except ValueError:
+                print("Insufficient data for this model. Skipping this slice.")
+
+    def run_disjoint(self, model, include_gaps=True, **kwargs):
+        """
+        Run the longitudinal model as initialized.
+
+        Arguments:
+            model (vigipy model): One of the supported vigipy models
+                                  from this module (i.e. gps, prr, etc)
+
+            include_gaps (bool): If a particular time slice has a
+                                 zero-count sum, still run the model?
+
+            kwargs: key word arguments can be added to this function call
+                    and they will be passed into the model at run time.
+
+        """
+        self.results = []
+        for count, (timestamp, subset) in zip(self.date_groups.sum()["count"], self.date_groups):
+            if count == 0 and not include_gaps:
+                continue
             sub_container = convert(subset)
             try:
                 da_results = model(sub_container, **kwargs)
@@ -55,4 +81,4 @@ class LongitudinalModel:
 
         """
         self.time_unit = time_unit
-        self.date_groups = self.data.resample(time_unit, on="date").sum()["count"]
+        self.date_groups = self.data.resample(time_unit, on="date")